hubspot_api/requirements.txt
requests==2.31.0
python-dotenv==1.0.0
pyyaml==6.0.1
hubspot-api-client==8.0.0 